[shared DatePicker] Make disabled styles more consistent.
