--- app/src/shared/components/CalendarIcon/calendarIcon.pcss
@@ -1,10 +1,0 @@
-.root {
-  bottom: 0;
-  cursor: pointer;
-  margin: 9px 2px;
-  height: 24px;
-  width: 24px;
-  pointer-events: none;
-  position: absolute;
-  right: 8px;
-}
